# [optimize](string) optimize split_by_string and substring_index function 