# Proposed changes
Use SIMD stringsearcher and SIMD memcmp optimze split_by_string and substring_index function.
- split_by_string function has 32%~540% up
- substring_index function has 22%~46% up

Performance difference depends on the needle size and whether the needle is constant param.  And the longer the needle, the more performance improvement

(1) split_by_string 32%~540% up
| needle | before | after |  up | comment |
|--------|--------| --------|--------|--------|
| needle size 6, constant | 6.6s ![image](https://user-images.githubusercontent.com/67053339/230769281-12a4f773-37d5-4896-975f-6ceb3fa9141c.png) | 1.14s ![image](https://user-images.githubusercontent.com/67053339/230769690-4951f5d0-8825-4b9f-bd3f-12c9ee1978b1.png) |  540% | use StringSearcher  |
| needle size 6, vector | 6.62s ![image](https://user-images.githubusercontent.com/67053339/230769318-f700dda0-be3d-45e4-a179-28e2eabc0183.png) | 2.94s ![image](https://user-images.githubusercontent.com/67053339/230769887-ef2e6cc3-f6b4-4fae-a1b1-8f2b4b5718c0.png) | 120% | use SIMD memcmp |
|  needle size 1, constant | 4.08s ![image](https://user-images.githubusercontent.com/67053339/230769845-1bf20350-a7e2-49aa-8dbc-296c61348c99.png) | 1.32s ![image](https://user-images.githubusercontent.com/67053339/230769905-063f2490-2f48-4495-b8b7-75862fa3ab3a.png) | 200% | use StringSearcher  |
|  needle size 1, vector |  4.14s <img width="646" alt="image" src="https://user-images.githubusercontent.com/67053339/230772655-809d6e85-5777-40dd-9dff-d14982a56397.png"> | 3.14s <img width="655" alt="image" src="https://user-images.githubusercontent.com/67053339/230772563-76089c52-0d79-4d0d-a313-3bad85bd8c64.png">  | 32%  | use SIMD memcmp  |


(2) substring_index  22%~46% up
1. needle vector params are not supported.
2. when needle size 1, memchr is 6% faster than SIMD stringsearcher. 
3. when needle size greater than 1, stringsearcher is better than memmem.

| needle | before | after |  up | comment |
|--------|--------| --------|--------|--------|
| needle size 6, constant | 1.48s ![image](https://user-images.githubusercontent.com/67053339/230770643-bc04da6f-67f7-45b0-850c-05126fcd0a09.png) | 1.03s ![image](https://user-images.githubusercontent.com/67053339/230770587-93e985c9-40c2-4984-95ab-b5529c4b2eb4.png) |  46% |  use StringSearcher  |
|  needle size 2, constant | 1.27s ![image](https://user-images.githubusercontent.com/67053339/230770798-9e207d27-89b5-4857-aea4-103daa290d13.png) | 1.03s ![image](https://user-images.githubusercontent.com/67053339/230770754-9b0ef88f-8899-41c3-8978-2fc0c4d73bfd.png) | 22% | use StringSearcher  |



Issue Number: close #xxx

## Problem summary

Describe your changes.

## Checklist(Required)

* [ ] Does it affect the original behavior
* [ ] Has unit tests been added
* [ ] Has document been added or modified
* [ ] Does it need to update dependencies
* [ ] Is this PR support rollback (If NO, please explain WHY)

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...